petit-gato/pyproject.toml
[tool.poetry]
name = "petit-gato"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pyodbc = "^5.1.0"
faker = "^25.9.2"
azure-identity = "^1.16.0"
azure-storage-file-datalake = "^12.15.0"
python-dotenv = "^1.0.1"
azure-core = "^1.30.1"
sqlalchemy = "^2.0.30"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
blue = "^0.9.1"
pytest = "^8.2.2"
taskipy = "^1.12.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
